Stop segfault on exit of test script
There seems to be a rather hard-to-find bug in cairocffi or cffi itself
that leads to a segfault on exit.  It looks like something may be trying
to destroy the surface for the window during python's cleanup after the
cairo library has already been unloaded.

Dropping the reference to the surface explicitly before exiting prevents
this.  It shouldn't be necessary but I can't find a simple bugfix for
cairocffi!