Remove implicit load of Wayland protocol at module init time
I was unhappy that importing wayland.protocol would try to read
/usr/share/wayland/wayland.xml and fail if it was not valid.  It should
be possible for users of this library to specify alternative locations
for the Wayland protocol definition files, and even use more than one at
once.

This commit removes that behaviour and requires clients to specify the
location of the wayland.xml file that they want to use.  The
client.Display proxy class is constructed explicitly after the protocol
is loaded.  In theory, a client could now be written that connects to
multiple Wayland displays using different versions of wayland.xml for
each one.

TODO: a protocol loader with search paths and defaults!